+++ Assets/Scripts/ResourceProduction/Producer.cs
@@ -72,7 +72,9 @@ namespace ResourceProduction {
         private void ProduceAtStart () {
             if (!data.AutoClickerActive || ProgressSinceQuit.ElapsedTime < 1)
                 return;
-            var produce = data.GetActualProductionAmount (Level) * NumberOwned * Mathf.RoundToInt (ProgressSinceQuit.ElapsedTime / data.GetActualProductionTime (NumberOwned));
+            var produce = data.GetActualProductionAmount (Level) * (ulong)NumberOwned * 
+                          (ulong)Mathf.RoundToInt (ProgressSinceQuit.ElapsedTime / data.GetActualProductionTime (NumberOwned));
+            
             ProgressSinceQuit.ProducedAmount += produce;
             data.Resource.CurrentAmount += produce;
         }
@@ -81,7 +83,7 @@ namespace ResourceProduction {
             progressBar.FollowProductionTime (data.GetActualProductionTime (NumberOwned), timer);
             if (timer < data.GetActualProductionTime (NumberOwned)) return;
             progressBar.ResetProgressbar ();
-            data.Resource.CurrentAmount += data.GetActualProductionAmount (Level) * NumberOwned;
+            data.Resource.CurrentAmount += data.GetActualProductionAmount (Level) * (ulong)NumberOwned;
             timer -= data.GetActualProductionTime (NumberOwned);
             isProducing = false;
         }
